Fixed `availableBonusesTypes.test.ts`
