+++ ts/features/bonus/bonusVacanze/store/reducers/__tests__/availableBonusesTypes.test.ts
@@ -18,7 +18,7 @@ import {
   ID_CGN_TYPE
 } from "../../../utils/bonus";
 import { BonusAvailable } from "../../../../../../../definitions/content/BonusAvailable";
-import * as availableBonusSelectors from "../availableBonusesTypes";
+import * as bonus from "../../../utils/bonus";
 
 const bonusMockContent = {
   name: "Bonus Vacanze",
@@ -41,17 +41,15 @@ const mockBonus: BonusAvailable = {
 
 describe("availableBonusesTypes with FF enabled", () => {
   it("should return 2 bonuses available", () => {
-    jest
-      .spyOn(availableBonusSelectors, "mapBonusIdFeatureFlag")
-      .mockImplementation(
-        // eslint-disable-next-line sonarjs/no-identical-functions
-        () =>
-          new Map<number, boolean>([
-            [ID_BONUS_VACANZE_TYPE, false],
-            [ID_BPD_TYPE, true],
-            [ID_CGN_TYPE, true]
-          ])
-      );
+    jest.spyOn(bonus, "mapBonusIdFeatureFlag").mockImplementation(
+      // eslint-disable-next-line sonarjs/no-identical-functions
+      () =>
+        new Map<number, boolean>([
+          [ID_BONUS_VACANZE_TYPE, false],
+          [ID_BPD_TYPE, true],
+          [ID_CGN_TYPE, true]
+        ])
+    );
     expect(
       supportedAvailableBonusSelector.resultFunc(
         pot.some([...availableBonuses])
@@ -60,18 +58,16 @@ describe("availableBonusesTypes with FF enabled", () => {
   });
 
   it("should return 2 bonuses available if an experimental bonus with no FF is available", () => {
-    jest
-      .spyOn(availableBonusSelectors, "mapBonusIdFeatureFlag")
-      .mockImplementation(
-        // eslint-disable-next-line sonarjs/no-identical-functions
-        () =>
-          new Map<number, boolean>([
-            [ID_BONUS_VACANZE_TYPE, false],
-            [ID_BPD_TYPE, true],
-            [ID_CGN_TYPE, true],
-            [4, true]
-          ])
-      );
+    jest.spyOn(bonus, "mapBonusIdFeatureFlag").mockImplementation(
+      // eslint-disable-next-line sonarjs/no-identical-functions
+      () =>
+        new Map<number, boolean>([
+          [ID_BONUS_VACANZE_TYPE, false],
+          [ID_BPD_TYPE, true],
+          [ID_CGN_TYPE, true],
+          [4, true]
+        ])
+    );
     const bonuses: BonusesAvailable = [
       ...availableBonuses,
       {
@@ -85,18 +81,16 @@ describe("availableBonusesTypes with FF enabled", () => {
   });
 
   it("should return 3 bonuses available if an experimental bonus with no FF is available", () => {
-    jest
-      .spyOn(availableBonusSelectors, "mapBonusIdFeatureFlag")
-      .mockImplementation(
-        // eslint-disable-next-line sonarjs/no-identical-functions
-        () =>
-          new Map<number, boolean>([
-            [ID_BONUS_VACANZE_TYPE, false],
-            [ID_BPD_TYPE, true],
-            [ID_CGN_TYPE, true],
-            [4, true]
-          ])
-      );
+    jest.spyOn(bonus, "mapBonusIdFeatureFlag").mockImplementation(
+      // eslint-disable-next-line sonarjs/no-identical-functions
+      () =>
+        new Map<number, boolean>([
+          [ID_BONUS_VACANZE_TYPE, false],
+          [ID_BPD_TYPE, true],
+          [ID_CGN_TYPE, true],
+          [4, true]
+        ])
+    );
     const bonuses: BonusesAvailable = [
       ...availableBonuses,
       {
@@ -123,18 +117,16 @@ describe("availableBonusesTypes with FF enabled", () => {
   });
 
   it("should return the experimental bonus where the FF is ON", () => {
-    jest
-      .spyOn(availableBonusSelectors, "mapBonusIdFeatureFlag")
-      .mockImplementation(
-        // eslint-disable-next-line sonarjs/no-identical-functions
-        () =>
-          new Map<number, boolean>([
-            [ID_BONUS_VACANZE_TYPE, false],
-            [ID_BPD_TYPE, true],
-            [ID_CGN_TYPE, true],
-            [4, true]
-          ])
-      );
+    jest.spyOn(bonus, "mapBonusIdFeatureFlag").mockImplementation(
+      // eslint-disable-next-line sonarjs/no-identical-functions
+      () =>
+        new Map<number, boolean>([
+          [ID_BONUS_VACANZE_TYPE, false],
+          [ID_BPD_TYPE, true],
+          [ID_CGN_TYPE, true],
+          [4, true]
+        ])
+    );
     const visibility = BonusVisibilityEnum.experimental;
     const bonuses: BonusesAvailable = [
       { ...mockBonus, id_type: ID_BONUS_VACANZE_TYPE, visibility },
@@ -149,18 +141,16 @@ describe("availableBonusesTypes with FF enabled", () => {
   });
 
   it("should return the experimental bonus where the FF is ON", () => {
-    jest
-      .spyOn(availableBonusSelectors, "mapBonusIdFeatureFlag")
-      .mockImplementation(
-        // eslint-disable-next-line sonarjs/no-identical-functions
-        () =>
-          new Map<number, boolean>([
-            [ID_BONUS_VACANZE_TYPE, false],
-            [ID_BPD_TYPE, true],
-            [ID_CGN_TYPE, true],
-            [4, true]
-          ])
-      );
+    jest.spyOn(bonus, "mapBonusIdFeatureFlag").mockImplementation(
+      // eslint-disable-next-line sonarjs/no-identical-functions
+      () =>
+        new Map<number, boolean>([
+          [ID_BONUS_VACANZE_TYPE, false],
+          [ID_BPD_TYPE, true],
+          [ID_CGN_TYPE, true],
+          [4, true]
+        ])
+    );
     const visibility = BonusVisibilityEnum.experimental;
     const bonuses: BonusesAvailable = [
       {
