+++ ts/features/bonus/bonusVacanze/utils/bonus.ts
@@ -10,6 +10,7 @@ import {
 } from "../store/actions/bonusVacanze";
 import { BonusActivationProgressEnum } from "../store/reducers/activation";
 import { EligibilityRequestProgressEnum } from "../store/reducers/eligibility";
+import { bonusVacanzeEnabled, bpdEnabled } from "../../../../config";
 
 export const ID_BONUS_VACANZE_TYPE = 1;
 export const ID_BPD_TYPE = 2;
@@ -65,3 +66,15 @@ const allowedActivationStatus = new Set([
 export const isActivationResponseTrackable = (
   eligibility: BonusVacanzeActivationPayload
 ) => allowedActivationStatus.has(eligibility.status);
+
+/**
+ * Return a Map with the bonus ID as a key and
+ * a bool representing the relative feature flag
+ * status as a value.
+ */
+export const mapBonusIdFeatureFlag = () =>
+  new Map<number, boolean>([
+    [ID_BONUS_VACANZE_TYPE, bonusVacanzeEnabled],
+    [ID_BPD_TYPE, bpdEnabled],
+    [ID_CGN_TYPE, true]
+  ]);
